Get rid of SyncService
