+++ Aika/DataFlow/State.swift
@@ -43,7 +43,7 @@ struct AppState : RxStateType {
 	let authentication: Authentication
 	let uiApplication: UIApplication
 	let authenticationService: AuthenticationServiceType
-	let syncService: SynchronizationServiceType
+	let webService: WebServiceType
 	let repository: RepositoryType
 	let syncStatus: SynchronizationStatus
 	var badgeStyle: IconBadgeStyle { return UserDefaults.standard.iconBadgeStyle }
@@ -61,13 +61,13 @@ extension AppStateMutation {
 	func new(coordinator: ApplicationCoordinatorType? = nil,
 	         authentication: Authentication? = nil,
 	         syncStatus: SynchronizationStatus? = nil,
-	         syncService: SynchronizationServiceType? = nil,
+	         webService: WebServiceType? = nil,
 	         repository: RepositoryType? = nil) -> AppState {
 		return AppState(coordinator: coordinator ?? state.coordinator,
 		                authentication: authentication ?? state.authentication,
 		                uiApplication: state.uiApplication,
 		                authenticationService: state.authenticationService,
-		                syncService: syncService ?? state.syncService,
+		                webService: webService ?? state.webService,
 		                repository: repository ?? state.repository,
 		                syncStatus: syncStatus ?? state.syncStatus)
 	}
